Code review, I moved components to the appropriate directories and removed unnecessary imports
